+++ x2engine/protected/modules/users/models/User.php
@@ -62,7 +62,7 @@ class User extends CActiveRecord {
      * @var bool If true, grid views displaying models of this type will have their filter and
      *  sort settings saved in the database instead of in the session
      */
-    public $persistentGridSettings = false;
+    public $dbPersistentGridSettings = false;
 
     /**
      * Returns the static model of the specified AR class.
@@ -81,17 +81,17 @@ class User extends CActiveRecord {
 
     public function behaviors(){
         return array_merge(parent::behaviors(), array(
-                    'X2LinkableBehavior' => array(
-                        'class' => 'X2LinkableBehavior',
-                        'module' => 'users',
-                        'viewRoute' => '/profile',
-                    ),
-                    'ERememberFiltersBehavior' => array(
-                        'class' => 'application.components.ERememberFiltersBehavior',
-                        'defaults' => array(),
-                        'defaultStickOnClear' => false
-                    )
-                ));
+            'X2LinkableBehavior' => array(
+                'class' => 'X2LinkableBehavior',
+                'module' => 'users',
+                'viewRoute' => '/profile',
+            ),
+            'ERememberFiltersBehavior' => array(
+                'class' => 'application.components.ERememberFiltersBehavior',
+                'defaults' => array(),
+                'defaultStickOnClear' => false
+            )
+        ));
     }
 
     /**
@@ -123,7 +123,7 @@ class User extends CActiveRecord {
             array('lastUpdated', 'length', 'max' => 30),
             array('userKey', 'length', 'max' => 32, 'min' => 3),
             array('backgroundInfo', 'safe'),
-            array('username','in','not'=>true,'range'=>array('Guest','Anyone'),'message'=>Yii::t('users','The specified username is reserved for system usage.')),
+            array('username','in','not'=>true,'range'=>array('Guest','Anyone',Profile::GUEST_PROFILE_USERNAME),'message'=>Yii::t('users','The specified username is reserved for system usage.')),
             array('username', 'unique', 'allowEmpty' => false),
             array('userAlias', 'unique', 'allowEmpty' => false),
             array('userAlias', 'match', 'pattern' => '/^\s+$/', 'not' => true),
@@ -217,10 +217,13 @@ class User extends CActiveRecord {
             $socialItem->delete();
         }
 
+        X2CalendarPermissions::model()->deleteAllByAttributes (
+            array(), 'user_id=:userId OR other_user_id=:userId', array (':userId' => $this->id)
+        );
+
         // delete profile
         $prof=Profile::model()->findByAttributes(array('username'=>$this->username));
-        if ($prof)
-            $prof->delete();
+        if ($prof) $prof->delete();
 
         // delete associated events
         Yii::app()->db->createCommand()
@@ -467,7 +470,7 @@ class User extends CActiveRecord {
                 array_pop($recentItems);  //remove the oldest ones
             }
             $userRecord->setAttribute('recentItems', implode(',', $recentItems));
-            $userRecord->save();
+            $userRecord->update();
         }
     }
